+++ src/Modules/Kongresse.php
@@ -53,7 +53,7 @@ class Kongresse extends \Module
 		$typen = unserialize($this->kongresse->typ);
 
 		// Datensätze laden
-		$objKongresse = \Database::getInstance()->prepare('SELECT * FROM tl_kongresse WHERE jahr >= ? AND jahr <= ? ORDER BY jahr DESC, datum DESC')
+		$objKongresse = \Database::getInstance()->prepare('SELECT * FROM tl_kongresse WHERE jahr >= ? AND jahr <= ? ORDER BY jahr DESC, datum_von DESC')
 		                                        ->execute($vonJahr, $bisJahr);
 
 		// Elo zuweisen
@@ -65,26 +65,21 @@ class Kongresse extends \Module
 			{
 				// Links zusammensetzen
 				$links = '';
+				$link_broschuere = '';
+				$link_protokoll = '';
 				if($objKongresse->file_broschuere)
 				{
-					$links .= '<a href="'.$this->replaceInsertTags('{{file::'.$objKongresse->file_broschuere.'}}').'">Kongressbroschüre</a>';
+					$link_broschuere = '<a href="'.$this->replaceInsertTags('{{file::'.$objKongresse->file_broschuere.'}}').'" target="_blank"><img src="bundles/contaokongresse/images/buch.png" alt="Buch/Broschüre"></a>';
 				}
 				if($objKongresse->file_protokoll)
 				{
-					if($links) $links .= ' | ';
-					$links .= '<a href="'.$this->replaceInsertTags('{{file::'.$objKongresse->file_protokoll.'}}').'">Protokoll</a>';
-				}
-				if($objKongresse->url)
-				{
-					if($links) $links .= ' | ';
-					$links .= '<a href="'.$this->replaceInsertTags($objKongresse->url).'">Infoseite</a>';
+					$link_protokoll = '<a href="'.$this->replaceInsertTags('{{file::'.$objKongresse->file_protokoll.'}}').'" target="_blank"><img src="bundles/contaokongresse/images/protokoll.png" alt="Protokoll"></a>';
 				}
 
 				// Extra-Links ermitteln
 				$extra = unserialize($objKongresse->extra_links);
 				if($extra)
 				{
-					if($links) $links .= '<br>';
 					foreach($extra as $key => $value)
 					{
 						$links .= '<a href="'.$this->replaceInsertTags($value['url']).'">'.$value['text'].'</a>';
@@ -92,14 +87,30 @@ class Kongresse extends \Module
 					}
 				}
 
+				// Jahr und Ort modifizieren
+				if($objKongresse->url)
+				{
+					$href = $this->replaceInsertTags($objKongresse->url);
+					$target = $objKongresse->newWindow ? ' target="_blank"' : '';
+					$jahr = '<a href="'.$href.'"'.$target.'>'.$objKongresse->jahr.'</a>';
+					$ort = '<a href="'.$href.'"'.$target.'>'.$objKongresse->ort.'</a>';
+				}
+				else
+				{
+					$jahr = $objKongresse->jahr;
+					$ort = $objKongresse->ort;
+				}
+
 				// Datensatz in Templatevariable schreiben
 				$records[] = array
 				(
-					'jahr'  => $objKongresse->jahr,
-					'ort'   => $objKongresse->ort,
-					'datum' => $objKongresse->datum ? date('d.m.Y', $objKongresse->datum) : '',
-					'info'  => $objKongresse->info,
-					'links' => $links,
+					'jahr'       => $jahr,
+					'ort'        => $ort,
+					'datum'      => self::DatumVerschmelzen($objKongresse->datum_von, $objKongresse->datum_bis),
+					'info'       => $objKongresse->info,
+					'broschuere' => $link_broschuere,
+					'protokoll'  => $link_protokoll,
+					'links'      => $links,
 				);
 			}
 		}
@@ -107,6 +118,62 @@ class Kongresse extends \Module
 		// Template füllen
 		$objTemplate = new \FrontendTemplate($this->strTemplate);
 		$this->Template->records = $records;
+	}
+
+	function DatumVerschmelzen($von, $bis)
+	{
+		// Starttag und Endetag vergleichen
+		if($von && $bis) 
+		{
+		  $start[0] = date("d",$von); // Starttag
+		  $start[1] = date("m",$von); // Startmonat
+		  $start[2] = date("Y",$von); // Startjahr
+		  $ende[0] = date("d",$bis); // Endetag
+		  $ende[1] = date("m",$bis); // Endemonat
+		  $ende[2] = date("Y",$bis); // Endejahr
+		  if($start[2] == $ende[2]) 
+		  {
+		    // gleiches Jahr
+		    $temp[0] = "";
+		    $temp[1] = $ende[2];
+		  }
+		  else
+		  {
+		    // unterschiedliches Jahr
+		    $temp[0] = $start[2];
+		    $temp[1] = $ende[2];
+		  }
+		  if($start[1] == $ende[1]) 
+		  {
+		    // gleicher Monat
+		    $temp[1] = $ende[1].".".$temp[1];
+		  }
+		  else
+		  {
+		    // unterschiedlicher Monat
+		    $temp[0] = $start[1].".".$temp[0];
+		    $temp[1] = $ende[1].".".$temp[1];
+		  }
+		  if($start[0] == $ende[0]) 
+		  {
+		    // gleicher Tag
+		    $temp[1] = $ende[0].".".$temp[1];
+		  }
+		  else
+		  {
+		    // unterschiedlicher Tag
+		    $temp[0] = $start[0].".".$temp[0];
+		    $temp[1] = $ende[0].".".$temp[1];
+		  }
+		  $anzeigetag = $temp[0]." - ".$temp[1];
+		}
+		elseif($von && !$bis)
+		{
+		  // Endetag ist nicht gesetzt
+		  $anzeigetag = date("d.m.Y",$von);
+		}    
+		else $anzeigetag = '';
+		return $anzeigetag;
 	}
 
 }