+++ src/Resources/contao/dca/tl_kongresse.php
@@ -27,14 +27,13 @@ $GLOBALS['TL_DCA']['tl_kongresse'] = array
 		'sorting' => array
 		(
 			'mode'                    => 2,
-			'fields'                  => array(),
-			'flag'                    => 1,
+			'fields'                  => array('jahr'),
+			'flag'                    => 12,
 			'panelLayout'             => 'filter;sort,search,limit',
 		),
 		'label' => array
 		(
-			// Das Feld aktiv wird vom label_callback überschrieben
-			'fields'                  => array('jahr','ort','datum','typ'),
+			'fields'                  => array('jahr','ort','datum_von','datum_bis','typ'),
 			'showColumns'             => true,
 			'format'                  => '%s',
 		),
@@ -81,14 +80,14 @@ $GLOBALS['TL_DCA']['tl_kongresse'] = array
 				'icon'                => 'visible.svg',
 				'attributes'          => 'onclick="Backend.getScrollOffset();return AjaxRequest.toggleVisibility(this,%s)"',
 				'button_callback'     => array('tl_kongresse', 'toggleIcon')
-			), 
+			),
 		)
 	),
 
 	// Paletten
 	'palettes' => array
 	(
-		'default'                     => '{congress_legend},typ,jahr,ort,datum,info;{files_legend},file_broschuere,file_protokoll,url;{extra_legend},extra_links;{aktiv_legend},aktiv'
+		'default'                     => '{congress_legend},typ,jahr,ort,datum_von,datum_bis,info;{files_legend},file_broschuere,file_protokoll,url,newWindow;{extra_legend},extra_links;{aktiv_legend},aktiv'
 	),
 
 	// Felder
@@ -114,7 +113,7 @@ $GLOBALS['TL_DCA']['tl_kongresse'] = array
 			'options'                 => &$GLOBALS['TL_LANG']['tl_kongresse']['typen'],
 			'eval'                    => array
 			(
-				'multiple'            => false, 
+				'multiple'            => false,
 				'tl_class'            => 'w50',
 				'includeBlankOption'  => true
 			),
@@ -143,15 +142,24 @@ $GLOBALS['TL_DCA']['tl_kongresse'] = array
 			'eval'                    => array('mandatory'=>false, 'maxlength'=>255, 'tl_class'=>'w50'),
 			'sql'                     => "varchar(255) NOT NULL default ''"
 		),
-		'datum' => array
+		'datum_von' => array
 		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_kongresse']['datum'],
+			'label'                   => &$GLOBALS['TL_LANG']['tl_kongresse']['datum_von'],
+			'exclude'                 => true,
+			'inputType'               => 'text',
+			'flag'                    => 5,
+			'eval'                    => array('rgxp'=>'date', 'datepicker'=>true, 'tl_class'=>'w50 wizard clr'),
+			'sql'                     => "varchar(11) NOT NULL default ''"
+		),
+		'datum_bis' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_kongresse']['datum_bis'],
 			'exclude'                 => true,
 			'inputType'               => 'text',
 			'flag'                    => 5,
 			'eval'                    => array('rgxp'=>'date', 'datepicker'=>true, 'tl_class'=>'w50 wizard'),
 			'sql'                     => "varchar(11) NOT NULL default ''"
-		), 
+		),
 		'info' => array
 		(
 			'label'                   => &$GLOBALS['TL_LANG']['tl_kongresse']['info'],
@@ -170,9 +178,9 @@ $GLOBALS['TL_DCA']['tl_kongresse'] = array
 			'inputType'               => 'fileTree',
 			'eval'                    => array
 			(
-				'filesOnly'           => true, 
-				'fieldType'           => 'radio', 
-				'mandatory'           => false, 
+				'filesOnly'           => true,
+				'fieldType'           => 'radio',
+				'mandatory'           => false,
 				'tl_class'            => 'w50'
 			),
 			'sql'                     => "binary(16) NULL"
@@ -184,9 +192,9 @@ $GLOBALS['TL_DCA']['tl_kongresse'] = array
 			'inputType'               => 'fileTree',
 			'eval'                    => array
 			(
-				'filesOnly'           => true, 
-				'fieldType'           => 'radio', 
-				'mandatory'           => false, 
+				'filesOnly'           => true,
+				'fieldType'           => 'radio',
+				'mandatory'           => false,
 				'tl_class'            => 'w50'
 			),
 			'sql'                     => "binary(16) NULL"
@@ -199,19 +207,34 @@ $GLOBALS['TL_DCA']['tl_kongresse'] = array
 			'inputType'               => 'text',
 			'eval'                    => array
 			(
-				'mandatory'           => false, 
-				'rgxp'                => 'url', 
-				'decodeEntities'      => true, 
-				'maxlength'           => 255, 
-				'dcaPicker'           => true, 
-				'addWizardClass'      => false, 
-				'tl_class'            => 'w50'
+				'mandatory'           => false,
+				'rgxp'                => 'url',
+				'decodeEntities'      => true,
+				'maxlength'           => 255,
+				'dcaPicker'           => true,
+				'addWizardClass'      => false,
+				'tl_class'            => 'w50 clr'
 			),
 			'sql'                     => "varchar(255) NOT NULL default ''"
 		),
+		'newWindow' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_kongresse']['newWindow'],
+			'exclude'                 => true,
+			'filter'                  => true,
+			'default'                 => true,
+			'inputType'               => 'checkbox',
+			'eval'                    => array
+			(
+				'mandatory'           => false,
+				'tl_class'            => 'w50',
+				'isBoolean'           => true,
+			),
+			'sql'                     => "char(1) NOT NULL default ''"
+		),
 		'extra_links' => array
 		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_kongresse']['url'],
+			'label'                   => &$GLOBALS['TL_LANG']['tl_kongresse']['extra_links'],
 			'exclude'                 => true,
 			'inputType'               => 'multiColumnWizard',
 			'eval'                    => array
@@ -222,9 +245,9 @@ $GLOBALS['TL_DCA']['tl_kongresse'] = array
 				(
 					'copy'            => 'system/themes/flexible/icons/copy.svg',
 					'delete'          => 'system/themes/flexible/icons/delete.svg',
-					'move'            => true,
-					'up'              => true,
-					'down'            => true
+					'move'            => 'system/themes/flexible/icons/move.svg',
+					'up'              => 'system/themes/flexible/icons/up.svg',
+					'down'            => 'system/themes/flexible/icons/down.svg'
 				),
 				'columnFields'        => array
 				(
@@ -236,12 +259,13 @@ $GLOBALS['TL_DCA']['tl_kongresse'] = array
 						'inputType'               => 'text',
 						'eval'                    => array
 						(
-							'mandatory'           => false, 
-							'rgxp'                => 'url', 
-							'decodeEntities'      => true, 
-							'maxlength'           => 255, 
-							'dcaPicker'           => true, 
-							'addWizardClass'      => false, 
+							'mandatory'           => false,
+							'rgxp'                => 'url',
+							'decodeEntities'      => true,
+							'maxlength'           => 255,
+							'dcaPicker'           => true,
+							'addWizardClass'      => false,
+							'style'               => 'width:90%;'
 						),
 					),
 					'text' => array
@@ -249,6 +273,23 @@ $GLOBALS['TL_DCA']['tl_kongresse'] = array
 						'label'                   => &$GLOBALS['TL_LANG']['tl_kongresse']['extra_links_text'],
 						'exclude'                 => true,
 						'inputType'               => 'text',
+						'eval'                    => array
+						(
+							'maxlength'           => 255,
+							'style'               => 'width:90%;'
+						),
+					),
+					'newWindow' => array
+					(
+						'label'                   => &$GLOBALS['TL_LANG']['tl_kongresse']['extra_links_newWindow'],
+						'exclude'                 => true,
+						'default'                 => true,
+						'inputType'               => 'checkbox',
+						'eval'                    => array
+						(
+							'mandatory'           => false,
+							'isBoolean'           => true,
+						),
 					),
 				)
 			),
@@ -276,26 +317,26 @@ class tl_kongresse extends Backend
 	public function toggleIcon($row, $href, $label, $title, $icon, $attributes)
 	{
 		$this->import('BackendUser', 'User');
-		
+
 		if (strlen($this->Input->get('tid')))
 		{
 			$this->toggleVisibility($this->Input->get('tid'), ($this->Input->get('state') == 0));
 			$this->redirect($this->getReferer());
 		}
-		
+
 		// Check permissions AFTER checking the tid, so hacking attempts are logged
 		if (!$this->User->isAdmin && !$this->User->hasAccess('tl_kongresse::aktiv', 'alexf'))
 		{
 			return '';
 		}
-		
+
 		$href .= '&amp;id='.$this->Input->get('id').'&amp;tid='.$row['id'].'&amp;state='.$row[''];
-		
+
 		if (!$row['aktiv'])
 		{
 			$icon = 'invisible.gif';
 		}
-		
+
 		return '<a href="'.$this->addToUrl($href).'" title="'.specialchars($title).'"'.$attributes.'>'.$this->generateImage($icon, $label).'</a> ';
 	}
 
@@ -308,9 +349,9 @@ class tl_kongresse extends Backend
 			$this->log('Not enough permissions to show/hide record ID "'.$intId.'"', 'tl_kongresse toggleVisibility', TL_ERROR);
 			$this->redirect('contao/main.php?act=error');
 		}
-		
+
 		$this->createInitialVersion('tl_kongresse', $intId);
-		
+
 		// Trigger the save_callback
 		if (is_array($GLOBALS['TL_DCA']['tl_kongresse']['fields']['aktiv']['save_callback']))
 		{
@@ -320,7 +361,7 @@ class tl_kongresse extends Backend
 				$blnPublished = $this->$callback[0]->$callback[1]($blnPublished, $this);
 			}
 		}
-		
+
 		// Update the database
 		$this->Database->prepare("UPDATE tl_kongresse SET tstamp=". time() .", aktiv='" . ($blnPublished ? '' : '1') . "' WHERE id=?")
 		     ->execute($intId);