+++ src/Resources/contao/languages/de/tl_kongresse.php
@@ -3,53 +3,35 @@
  * Buttons für Operationen
  */
 
-$GLOBALS['TL_LANG']['tl_kongresse']['categories'] = array('Kategorien', 'Kategorien für die Adressen bearbeiten');
-
-$GLOBALS['TL_LANG']['tl_kongresse']['import'][0] = 'CSV-Import';
-$GLOBALS['TL_LANG']['tl_kongresse']['import'][1] = 'Daten aus einer CSV-Datei importieren';
-
-$GLOBALS['TL_LANG']['tl_kongresse']['export'][0] = 'CSV-Export';
-$GLOBALS['TL_LANG']['tl_kongresse']['export'][1] = 'Daten als CSV-Datei exportieren';
-
-$GLOBALS['TL_LANG']['tl_kongresse']['new'][0] = 'Neue Adresse';
-$GLOBALS['TL_LANG']['tl_kongresse']['new'][1] = 'Neue Adresse anlegen';
-
-$GLOBALS['TL_LANG']['tl_kongresse']['edit'][0] = 'Adresse bearbeiten';
-$GLOBALS['TL_LANG']['tl_kongresse']['edit'][1] = 'Adresse %s bearbeiten';
-
-$GLOBALS['TL_LANG']['tl_kongresse']['copy'][0] = 'Adresse kopieren';
-$GLOBALS['TL_LANG']['tl_kongresse']['copy'][1] = 'Adresse %s kopieren';
-
-$GLOBALS['TL_LANG']['tl_kongresse']['delete'][0] = 'Adresse löschen';
-$GLOBALS['TL_LANG']['tl_kongresse']['delete'][1] = 'Adresse %s löschen';
-
-$GLOBALS['TL_LANG']['tl_kongresse']['toggle'][0] = 'Adresse aktivieren/deaktivieren';
-$GLOBALS['TL_LANG']['tl_kongresse']['toggle'][1] = 'Adresse %s aktivieren/deaktivieren';
-
-$GLOBALS['TL_LANG']['tl_kongresse']['show'][0] = 'Adressdetails anzeigen';
-$GLOBALS['TL_LANG']['tl_kongresse']['show'][1] = 'Details der Adresse %s anzeigen';
+$GLOBALS['TL_LANG']['tl_kongresse']['new'] = array('Neue Veranstaltung', 'Neue Veranstaltung anlegen');
+$GLOBALS['TL_LANG']['tl_kongresse']['edit'] = array('Veranstaltung bearbeiten', 'Veranstaltung %s bearbeiten');
+$GLOBALS['TL_LANG']['tl_kongresse']['copy'] = array('Veranstaltung kopieren', 'Veranstaltung %s kopieren');
+$GLOBALS['TL_LANG']['tl_kongresse']['delete'] = array('Veranstaltung löschen', 'Veranstaltung %s löschen');
+$GLOBALS['TL_LANG']['tl_kongresse']['toggle'] = array('Veranstaltung aktivieren/deaktivieren', 'Veranstaltung %s aktivieren/deaktivieren');
+$GLOBALS['TL_LANG']['tl_kongresse']['show'] = array('Veranstaltungsdetails anzeigen', 'Details der Veranstaltung %s anzeigen');
 
 /**
  * Backend-Modul: Übersetzungen im Eingabeformular
  */
-$GLOBALS['TL_LANG']['tl_kongresse']['id'] = array('ID', 'ID des Datensatzes');
-$GLOBALS['TL_LANG']['tl_kongresse']['tstamp'] = array('Letzte Änderung', 'Zeitstempel der letzten Änderung');
-
 $GLOBALS['TL_LANG']['tl_kongresse']['congress_legend'] = 'Veranstaltung';
 $GLOBALS['TL_LANG']['tl_kongresse']['typ'] = array('Art', 'Art der Veranstaltung');
 $GLOBALS['TL_LANG']['tl_kongresse']['jahr'] = array('Jahr', 'Jahr der Veranstaltung');
 $GLOBALS['TL_LANG']['tl_kongresse']['ort'] = array('Ort', 'Ort der Veranstaltung');
-$GLOBALS['TL_LANG']['tl_kongresse']['datum'] = array('Datum', 'Datum im Format TT.MM.JJJJ oder leerlassen');
+$GLOBALS['TL_LANG']['tl_kongresse']['datum_von'] = array('Von Datum', 'Von-Datum im Format TT.MM.JJJJ oder leerlassen');
+$GLOBALS['TL_LANG']['tl_kongresse']['datum_bis'] = array('Bis Datum', 'Bis-Datum im Format TT.MM.JJJJ oder leerlassen');
 $GLOBALS['TL_LANG']['tl_kongresse']['info'] = array('Information', 'Zusätzliche Informationen, die angezeigt werden sollen');
 
 $GLOBALS['TL_LANG']['tl_kongresse']['files_legend'] = 'Standard-Downloads und Veranstaltungsseite';
-$GLOBALS['TL_LANG']['tl_kongresse']['file_broschuere'] = array('Kongressbroschüre', 'Kongressbroschüre der Veranstaltung');
+$GLOBALS['TL_LANG']['tl_kongresse']['file_broschuere'] = array('Kongressbuch/-broschüre', 'Kongressbuch oder -broschüre der Veranstaltung');
 $GLOBALS['TL_LANG']['tl_kongresse']['file_protokoll'] = array('Protokoll', 'Protokoll der Veranstaltung');
 $GLOBALS['TL_LANG']['tl_kongresse']['url'] = array('Link', 'Link zur Veranstaltungsseite');
+$GLOBALS['TL_LANG']['tl_kongresse']['newWindow'] = array('Neues Fenster', 'Link im neuen Fenster öffnen');
 
 $GLOBALS['TL_LANG']['tl_kongresse']['extra_legend'] = 'Weitere Links';
-$GLOBALS['TL_LANG']['tl_kongresse']['extra_links_url'] = array('URL', 'Link zur Seite oder Datei');
-$GLOBALS['TL_LANG']['tl_kongresse']['extra_links_text'] = array('Linktext', 'Linktext der URL (optional)');
+$GLOBALS['TL_LANG']['tl_kongresse']['extra_links'] = array('Links', '');
+$GLOBALS['TL_LANG']['tl_kongresse']['extra_links_url'] = array('URL', '');
+$GLOBALS['TL_LANG']['tl_kongresse']['extra_links_text'] = array('Linktext (optional)', '');
+$GLOBALS['TL_LANG']['tl_kongresse']['extra_links_newWindow'] = array('Neues Fenster', '');
 
 $GLOBALS['TL_LANG']['tl_kongresse']['aktiv_legend'] = 'Veröffentlichung';
 $GLOBALS['TL_LANG']['tl_kongresse']['aktiv'] = array('Veröffentlicht', 'Datensatz veröffentlichen');
@@ -58,6 +40,6 @@ $GLOBALS['TL_LANG']['tl_kongresse']['typen'] = array
 (
 	'oK' => 'Ordentlicher Kongress',
 	'aK' => 'Außerordentlicher Kongress',
-	'Ha' => 'Hauptausschuss'
+	'Ha' => 'Hauptausschuss',
+	'Ko' => 'Kommissionstagung'
 );
-