+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Kongresse und Hauptausschüsse Changelog
 
+## Version 1.1.0 (2020-09-29)
+
+* Add: TODO.md
+* Datensatz bearbeiten: Kongressdatum von-bis statt nur von
+* Add: Veranstaltungstyp Kommissionstagung
+* Add: Links optional im neuen Fenster öffnen
+* Datensatz bearbeiten: MCW-Ansicht korrigieren
+* Add: Icons für Broschüre und Protokoll
+* Template-Ausgabe: CSS-Klassen bei den Tabellenspalten
+* Template: Ort/Jahr verlinken mit Infoseite, dafür Infoseite bei den Links entfernen
+
 ## Version 1.0.0 (2020-09-24)
 
 * DCA ausgebaut